+++ ios/CustomerioReactnative.m
@@ -23,6 +23,8 @@ RCT_EXTERN_METHOD(screen: (nonnull NSString *) name
 
 RCT_EXTERN_METHOD(registerDeviceToken : (nonnull NSString *) token)
 
+RCT_EXTERN_METHOD(deleteDeviceToken)
+
 RCT_EXTERN_METHOD(showPromptForPushNotifications: (NSDictionary *) options
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)